// frontend/src/gen/types/EquipmentOut.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { EquipmentPhotoOut } from "./EquipmentPhotoOut.ts";
import type { EquipmentStatus } from "./EquipmentStatus.ts";
import type { ServiceRecordOut } from "./ServiceRecordOut.ts";

/**
 * EquipmentOut
 */
export type EquipmentOut = {
  /**
   * @type integer
   */
  id: number;
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  make_model: string;
  /**
   * @type string
   */
  serial_number: string;
  purchase_date: string | null;
  /**
   * @type string
   */
  notes: string;
  /**
   * @type string
   */
  status: EquipmentStatus;
  service_interval_days: number | null;
  last_serviced_on: string | null;
  next_service_due: string | null;
  /**
   * @type string
   */
  service_contact_name: string;
  /**
   * @type string
   */
  service_contact_phone: string;
  /**
   * @type string
   */
  service_contact_email: string;
  /**
   * @type string
   */
  service_contact_notes: string;
  /**
   * @type array
   */
  photos: EquipmentPhotoOut[];
  /**
   * @type array
   */
  service_records: ServiceRecordOut[];
};